API/app/requirements.txt
asgiref==3.10.0
attrs==25.4.0
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-standardized-errors==0.15.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
